refactor: move generated types to a separate folder
