+++ application-templates/starter-typescript/src/components/channel-details/types.ts
@@ -1,4 +1,4 @@
-import type { TChannelRole } from '../../../@types/generated/ctp';
+import type { TChannelRole } from '../../../types/generated/ctp';
 
 export type FormValues = {
   key: string;
